+++ src/client/components/SignupDialog.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Dialog, FlatButton, TextField, IconButton } from 'material-ui';
-import '../../client/styles/style.scss';
+import '../../../public/styles/style.scss';
 import IP from '../../../config/config.js';
 import { validateLogin, validatePassword, validateEmail } from './common.js';
