+++ src/client/components/helpers/marker_styles.js
