+++ src/client/components/preferences-list/CategorySlider.jsx
